backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
razorpay>=1.4
python-dotenv>=1.0
